Actables are now implemented
Reactables are checked each Update, Interactables are called only when the interact button is pushed.
Change to the PlayerController, so that Sit and Interact are only called once with each push of the button
Pebble kicking now in the game. For now it initiates when pebbles are colliding.
Pebbles now disappear when going over the edge of a tile, just like rabbits. Still no good solution to multi-tile tracking
All changes should be immidiately available in all scenes